Fix -Wsign-compare warnings in EXIF

+++ ext/exif/exif.c
@@ -2196,7 +2196,7 @@ static void exif_iif_add_value(image_info_type *image_info, int section_index, c
 
 	switch (format) {
 		case TAG_FMT_STRING:
-			if (length > value_len) {
+			if ((size_t)length > value_len) {
 				exif_error_docref("exif_iif_add_value" EXIFERR_CC, image_info, E_WARNING, "length > value_len: %d > %zu", length, value_len);
 				value = NULL;
 			}
@@ -2223,7 +2223,7 @@ static void exif_iif_add_value(image_info_type *image_info, int section_index, c
 			if (!length)
 				break;
 		case TAG_FMT_UNDEFINED:
-			if (length > value_len) {
+			if ((size_t)length > value_len) {
 				exif_error_docref("exif_iif_add_value" EXIFERR_CC, image_info, E_WARNING, "length > value_len: %d > %zu", length, value_len);
 				value = NULL;
 			}
